+++ src/app/screens/PaymentsHistory/components/PaymentDetail/styles.module.scss
