+++ src/config/segba/config.js
@@ -2,7 +2,7 @@ import i18 from 'i18next';
 
 import Colors from './scss/_colors.scss';
 
-export const columnsSegba = [
+const columns = [
   {
     name: i18.t('Payments:PaymentDetail:datetime'),
     key: 'datetime',
@@ -34,5 +34,6 @@ export default {
   },
   billsHistory: {
     enabled: false
-  }
+  },
+  columns
 };
